requirements.txt
feedparser==6.0.11
trafilatura==1.7.0
requests==2.33.0
beautifulsoup4==4.12.3
lxml==5.3.0
google-cloud-texttospeech==2.16.5
google-cloud-bigquery==3.25.0
pydub==0.25.1
mutagen==1.47.0
python-dotenv==1.0.1
PyYAML==6.0.2
internetarchive==5.5.1
feedgen==0.9.0
defusedxml==0.7.1
audioop-lts==0.2.1; python_version >= "3.13"
